requirements.txt
# Web Framework
fastapi
uvicorn
httpx
redis

# Cloud (AWS SDK)
boto3
python-dotenv

# NLP
spacy
nltk
requests

# (Optional) Hugging Face transformers for later
transformers
torch


#new req:
python-multipart
pypdf
pdfplumber
python-docx
sentence-transformers
numpy
scikit-learn
redis
rq  # or celery[redis]
pydantic
loguru



regex

